Time: 0 ms, Memory: 8.6 MB - LeetHub

+++ 0492-76-2027-minimum-moves-to-convert-string/0492-76-2027-minimum-moves-to-convert-string.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minimumMoves(string s) {
-        long moves = 0;
+        int moves = 0;
         int i=0;
             while(i<s.size()){
                 if(s[i] == 'X'){
